libxl: Wait for Qemu spawned by xl devd to be ready before continue
The toolstack shouldn't unpause guest until "non local" device-model
(currently Qdisk backend) is up and running. As guest checks
the availability of the Virtio devices (and discovers them) at
the early stage and QEMU may not be completely initialized yet by
that time. If that race happens the Virtio devices that Qemu is
intended to emulate will never show up in the guest. This race has
never been observed for Qemu running in the toolstack domain,
maybe because of the QMP support that toolstack uses for the *local*
device-model.

So, wait it to be ready by periodically checking the presence of
"state=running" in Xenstore. Done this right after adding Virtio devices
which in turn should trigger xl devd in driver domain to spawn Qemu.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>