libxl/arm: Rework virtio-pci support for dynamic PCI host allocation
In order to enable more use-cases such as having multiple
device-models (Qemu) running in different backend domains which provide
virtio-pci devices for the same guest, we allocate and expose one
PCI host bridge for every virtio backend domain for that guest.

For that purpose, reserve separate virtio-pci resources (memory and irqs)
up to 8 possible PCI hosts (to be aligned with MAX_NR_IOREQ_SERVERS)
and allocate a host per backend domain. The PCI host details including
its host_id to be written to dedicated Xenstore node for the device-model
to retrieve.

Also do not use vPCI resources for virtio-pci at all and restore vPCI
device-tree node as it was before the introduction of basic virtio-pci
support.

With that we won't break the PCI passthrough if both features are enabled
for the same guest.

For the upstream this should be squashed with:
"libxl/arm: Add basic virtio-pci support"
and subsequent commits for the toolstack.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>